requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.3
tenacity==8.2.3
python-dateutil==2.9.0.post0
